requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
pydantic-settings>=2.0.0
httpx>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.9
PyMuPDF>=1.24.0
python-docx>=1.1.0
openpyxl>=3.1.0
ddgs>=9.0.0
apscheduler>=3.10.0
